feat: use sonic decode
