+++ ast/decode.go
@@ -19,7 +19,6 @@ package ast
 import (
     `encoding/base64`
     `runtime`
-    `strconv`
     `unsafe`
 
     `github.com/bytedance/sonic/internal/native/types`
@@ -94,24 +93,6 @@ func decodeFalse(src string, pos int) (ret int) {
     return -int(types.ERR_INVALID_CHAR)
 }
 
-//go:nocheckptr
-func decodeString(src string, pos int) (ret int, v string) {
-    ret, ep := skipString(src, pos)
-    if ep == -1 {
-        (*rt.GoString)(unsafe.Pointer(&v)).Ptr = rt.IndexChar(src, pos+1)
-        (*rt.GoString)(unsafe.Pointer(&v)).Len = ret - pos - 2
-        return ret, v
-    }
-
-    vv, ok := unquoteBytes(rt.Str2Mem(src[pos:ret]))
-    if !ok {
-        return -int(types.ERR_INVALID_CHAR), ""
-    }
-
-    runtime.KeepAlive(src)
-    return ret, rt.Mem2Str(vv)
-}
-
 func decodeBinary(src string, pos int) (ret int, v []byte) {
     var vv string
     ret, vv = decodeString(src, pos)
@@ -130,94 +111,8 @@ func isDigit(c byte) bool {
     return c >= '0' && c <= '9'
 }
 
-//go:nocheckptr
-func decodeInt64(src string, pos int) (ret int, v int64, err error) {
-    sp := uintptr(rt.IndexChar(src, pos))
-    ss := uintptr(sp)
-    se := uintptr(rt.IndexChar(src, len(src)))
-    if uintptr(sp) >= se {
-        return -int(types.ERR_EOF), 0, nil
-    }
-
-    if c := *(*byte)(unsafe.Pointer(sp)); c == '-' {
-        sp += 1
-    }
-    if sp == se {
-        return -int(types.ERR_EOF), 0, nil
-    }
-
-    for ; sp < se; sp += uintptr(1) {
-        if !isDigit(*(*byte)(unsafe.Pointer(sp))) {
-            break
-        }
-    }
-
-    if sp < se {
-        if c := *(*byte)(unsafe.Pointer(sp)); c == '.' || c == 'e' || c == 'E' {
-            return -int(types.ERR_INVALID_NUMBER_FMT), 0, nil
-        }
-    }
-
-    var vv string
-    ret = int(uintptr(sp) - uintptr((*rt.GoString)(unsafe.Pointer(&src)).Ptr))
-    (*rt.GoString)(unsafe.Pointer(&vv)).Ptr = unsafe.Pointer(ss)
-    (*rt.GoString)(unsafe.Pointer(&vv)).Len = ret - pos
-
-    v, err = strconv.ParseInt(vv, 10, 64)
-    if err != nil {
-        //NOTICE: allow overflow here
-        if err.(*strconv.NumError).Err == strconv.ErrRange {
-            return ret, 0, err
-        }
-        return -int(types.ERR_INVALID_CHAR), 0, err
-    }
-
-    runtime.KeepAlive(src)
-    return ret, v, nil
-}
-
 func isNumberChars(c byte) bool {
     return (c >= '0' && c <= '9') || c == '+' || c == '-' || c == 'e' || c == 'E' || c == '.'
-}
-
-//go:nocheckptr
-func decodeFloat64(src string, pos int) (ret int, v float64, err error) {
-    sp := uintptr(rt.IndexChar(src, pos))
-    ss := uintptr(sp)
-    se := uintptr(rt.IndexChar(src, len(src)))
-    if uintptr(sp) >= se {
-        return -int(types.ERR_EOF), 0, nil
-    }
-
-    if c := *(*byte)(unsafe.Pointer(sp)); c == '-' {
-        sp += 1
-    }
-    if sp == se {
-        return -int(types.ERR_EOF), 0, nil
-    }
-
-    for ; sp < se; sp += uintptr(1) {
-        if !isNumberChars(*(*byte)(unsafe.Pointer(sp))) {
-            break
-        }
-    }
-
-    var vv string
-    ret = int(uintptr(sp) - uintptr((*rt.GoString)(unsafe.Pointer(&src)).Ptr))
-    (*rt.GoString)(unsafe.Pointer(&vv)).Ptr = unsafe.Pointer(ss)
-    (*rt.GoString)(unsafe.Pointer(&vv)).Len = ret - pos
-
-    v, err = strconv.ParseFloat(vv, 64)
-    if err != nil {
-        //NOTICE: allow overflow here
-        if err.(*strconv.NumError).Err == strconv.ErrRange {
-            return ret, 0, err
-        }
-        return -int(types.ERR_INVALID_CHAR), 0, err
-    }
-
-    runtime.KeepAlive(src)
-    return ret, v, nil
 }
 
 func decodeValue(src string, pos int, skipnum bool) (ret int, v types.JsonState) {
